Change some resource group variables type to int64 to avoid overflow.
Change of totalQueued and totalExecuted from int to int64 to avoid overflow after long running.

+++ src/backend/utils/resgroup/resgroup.c
@@ -139,8 +139,8 @@ struct ResGroupData
 
 	volatile int	nRunning;			/* number of running trans */
 	volatile int	nRunningBypassed;	/* number of running trans in bypass mode */
-	int				totalExecuted;		/* total number of executed trans */
-	int				totalQueued;		/* total number of queued trans	*/
+	int64			totalExecuted;		/* total number of executed trans */
+	int64			totalQueued;		/* total number of queued trans	*/
 	int64			totalQueuedTimeMs;	/* total queue time, in milliseconds */
 	PROC_QUEUE		waitProcs;			/* list of PGPROC objects waiting on this group */
